controller/requirements.txt
names-generator==0.1.0
python-dotenv==0.21.0
pyfiglet==0.8.post1
requests==2.28.1
tabulate==0.9.0
termcolor==2.2.0
urllib3==1.26.13
